tmux.conf: set "terminal-overrides" to avoid strange char
Strange characters '???112' may appear when exited the tmux session,
which is caused by the terminal that dose not properly handle the Cr
(color reset) control sequence.  So use "terminal-overrides" to override
this.

Environment: SSH to a DragonFly BSD with Xfce4 terminal. (other
terminals, e.g., urxvt and konsole, are OK)