zsh/gpg: ignore agent error in preexec hook
For a NFS-shared-home cluster, once login to a node and start the
gpg-agnet, then the original gpg-agent on the master will quit, which
causes the annoying agent connection error.

Simply ignore the error and do not try to launch the agent, which is of
little use for the above cases.